Move signature code into it's own module
